+++ asynctnt/__init__.py
@@ -1,3 +1,4 @@
+from .const import Transport
 from .connection import Connection, connect
 from .iproto.protocol import (
     Iterator, Response, TarantoolTuple, PushIterator,