+++ asynctnt/const.py
@@ -1,0 +1,5 @@
+import enum
+
+class Transport(enum.IntEnum):
+    DEFAULT = 1
+    SSL = 2